feat: split config driver methods (#414)
